requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-ltr390==1.1.17
adafruit-circuitpython-register==1.10.1
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.1
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
pyftdi==0.56.0
pyserial==3.5
pytz==2024.2
pyusb==1.2.1
RPi.bme280==0.2.4
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
smbus2==0.5.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
